Recoded Wall Spawning system
The walls now handle the choice between being a wall and a door inside themselves, instead of this being handled by the floor of the room. This allows the system to be much better at detecting if a wall should be a wall or if it should be a door.